Improve cubic drawing, remove useless repo, better code format
