+++ pom.xml
@@ -67,12 +67,5 @@
           <version>1.13-R0.1-SNAPSHOT</version>
           <scope>provided</scope>
       </dependency>
-      <dependency>
-          <groupId>cn.suxiaolin</groupId>
-          <artifactId>suCore</artifactId>
-          <version>1.0.0</version>
-          <scope>system</scope>
-          <systemPath>${project.basedir}/libs/suCore-1.0.0.jar</systemPath>
-      </dependency>
   </dependencies>
 </project>
