+++ src/main/java/cn/suxiaolin/subuildingpacking/util/RandomID.java
@@ -3,29 +3,19 @@ package cn.suxiaolin.subuildingpacking.util;
 import java.util.Random;
 
 public class RandomID {
+
+    private static final String validChars = "0123456789abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ_/";
+    private static final Random random = new Random();
+
     private static char getRandomChar() {
-        Random random = new Random();
-        int type = random.nextInt(3); // 0: 数字, 1: 小写字母, 2: 大写字母
-        switch (type) {
-            case 0:
-                return (char) ('0' + random.nextInt(10));
-            case 1:
-                return (char) ('a' + random.nextInt(26));
-            case 2:
-                return (char) ('A' + random.nextInt(26));
-            default:
-                throw new IllegalStateException("Unexpected value: " + type);
-        }
+        return validChars.charAt(random.nextInt(64));
     }
 
     public static String generateRandomID() {
         StringBuilder codeBuilder = new StringBuilder();
-
         for (int i = 0; i < 5; i++) {
-            char randomChar = getRandomChar();
-            codeBuilder.append(randomChar);
+            codeBuilder.append(getRandomChar());
         }
-
         return codeBuilder.toString();
     }
 }
